requirements.txt
# ── Core ML ───────────────────────────────────────────────────────────────────
scikit-learn>=1.5.0
xgboost>=2.1.0
numpy>=1.26.0
pandas>=2.2.0
joblib>=1.4.0
scipy>=1.13.0

# ── Optimisation ──────────────────────────────────────────────────────────────
cvxpy>=1.5.0

# ── Deep Learning (LSTM for SoC prediction) ───────────────────────────────────
# torch>=2.3.0   # Optional — uncomment to enable LSTM model

# ── Backend ───────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9

# ── Async / Task Queue ────────────────────────────────────────────────────────
celery[redis]>=5.4.0
redis>=5.0.7

# ── Database ──────────────────────────────────────────────────────────────────
influxdb-client>=1.44.0
asyncpg>=0.29.0
sqlalchemy>=2.0.31

# ── Messaging ─────────────────────────────────────────────────────────────────
kafka-python>=2.0.2
paho-mqtt>=2.1.0

# ── HTTP / Weather API ────────────────────────────────────────────────────────
requests>=2.32.0
httpx>=0.27.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1
structlog>=24.2.0
tenacity>=8.5.0

# ── Development / Testing ─────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.7
httpx>=0.27.0       # for FastAPI TestClient
black>=24.4.2
ruff>=0.5.0
